requirements.txt
dash>=2.14.0
dash-core-components>=2.0.0
dash-bootstrap-components>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
gunicorn>=21.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Trading APIs
alpaca-trade-api>=2.3.0
oandapyV20>=0.7.0
yfinance>=0.2.0
v20>=0.30.0

# Technical Analysis
pandas_ta>=0.3.0
numpy-financial>=1.1.0

# Data & Calendar
pandas_market_calendars>=4.3.0
pandas-datareader>=0.10.0

# Utilities
matplotlib>=3.8.0
python-dateutil>=2.8.0
pytz>=2023.3

# Logging
python-json-logger>=2.0.0
alpaca-trade-api
alpaca-py
apscheduler